@@ -2,7 +2,7 @@
     "name": "linc/nitro-porter",
     "type": "project",
     "description": "The only multi-platform community migration tool. Free your forum!",
-    "license": "GPL-2.0-or-later",
+    "license": "AGPL-3.0-only",
     "authors": [
         {
             "name": "Lincoln Russell",
@@ -31,23 +31,24 @@
         "bin/porter"
     ],
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.2",
         "ext-gd": "*",
         "ext-json": "*",
         "ext-pdo": "*",
         "ext-zlib": "*",
-        "illuminate/database": "^8.0",
-        "monolog/monolog": "^2",
-        "s9e/text-formatter": "^2",
-        "nadar/quill-delta-parser": "^2"
+        "illuminate/database": "^11",
+        "monolog/monolog": "^3",
+        "adhocore/cli": "^1",
+        "s9e/text-formatter": "2.15.*",
+        "nadar/quill-delta-parser": "^3"
     },
     "require-dev": {
-        "phing/phing": "^2.0",
-        "robmorgan/phinx": "^0.12",
-        "fakerphp/faker": "^1.16",
+        "phing/phing": "^2",
+        "robmorgan/phinx": "^0",
+        "fakerphp/faker": "^1",
         "squizlabs/php_codesniffer": "^3",
         "phpunit/phpunit": "^9",
-        "phpstan/phpstan": "^1",
+        "phpstan/phpstan": "^2",
         "marcocesarato/php-conventional-changelog": "^1"
     },
     "autoload": {
