@@ -3,7 +3,6 @@
 /**
  * ASP Playground exporter tool
  *
- * @license http://opensource.org/licenses/gpl-2.0.php GNU GPL2
  * @author  Lincoln Russell, lincolnwebs.com
  */
 
@@ -22,11 +21,17 @@ class AspPlayground extends Source
         ],
         'features' => [
             'Users' => 1,
-            'Passwords' => 0,
             'Categories' => 1,
             'Discussions' => 1,
             'Comments' => 1,
+            'Bookmarks' => 1,
+            'Polls' => 0, // Challenges noted inline below.
+            'PrivateMessages' => 0, // Don't appear to be threaded in a rational way (see table PMsg).
         ]
+    ];
+
+    protected const FLAGS = [
+        'hasDiscussionBody' => false,
     ];
 
     /**
@@ -35,11 +40,14 @@ class AspPlayground extends Source
     public function run($ex)
     {
         $this->users($ex);
-        $this->roles($ex);
         $this->signatures($ex);
+
         $this->categories($ex);
+
         $this->discussions($ex);
         $this->comments($ex);
+        $this->bookmarks($ex);
+        $this->polls($ex);
     }
 
     /**
@@ -47,37 +55,23 @@ class AspPlayground extends Source
      */
     protected function users(ExportModel $ex): void
     {
-        $user_Map = array(
+        $map = [
             'Mem' => 'UserID',
             'Login' => 'Name',
             'Email' => 'Email',
             'Userpass' => 'Password',
             'totalPosts' => 'CountComments',
-            'ip' => 'LastIPAddress',
             'banned' => 'Banned',
             'dateSignUp' => 'DateInserted',
             'lastLogin' => 'DateLastActive',
-        );
+            'location' => 'Location',
+        ];
         $ex->export(
             'User',
-            "
-         select m.*,
-            'Text' as HashMethod
-         from :_Members m;",
-            $user_Map
+            "select m.*, 'Text' as HashMethod
+                from :_Members m",
+            $map
         );
-    }
-
-    /**
-     * @param ExportModel $ex
-     */
-    protected function roles(ExportModel $ex): void
-    {
-        // Make everyone a member since there's no used roles.
-        $userRole_Map = array(
-            'Mem' => 'UserID'
-        );
-        $ex->export('UserRole', 'select Mem, 8 as RoleID from :_Members', $userRole_Map);
     }
 
     /**
@@ -87,22 +81,15 @@ class AspPlayground extends Source
     {
         $ex->export(
             'UserMeta',
-            "
-         select
-            Mem,
-            'Plugin.Signatures.Sig' as `Name`,
-            signature as `Value`
-         from :_Members
-         where signature <> ''
+            "select Mem, 'Plugin.Signatures.Sig' as `Name`, signature as `Value`
+            from :_Members
+            where signature <> ''
 
-         union all
+            union all
 
-         select
-            Mem,
-            'Plugin.Signatures.Format' as `Name`,
-            'BBCode' as `Value`
-         from :_Members
-         where signature <> '';"
+            select Mem, 'Plugin.Signatures.Format' as `Name`, 'BBCode' as `Value`
+            from :_Members
+            where signature <> ''"
         );
     }
 
@@ -111,18 +98,22 @@ class AspPlayground extends Source
      */
     protected function categories(ExportModel $ex): void
     {
-        $category_Map = array(
+        $map = [
             'ForumID' => 'CategoryID',
             'ForumTitle' => 'Name',
             'ForumDesc' => 'Description',
             'Sort' => 'Sort',
-            'lastModTime' => 'DateUpdated'
-        );
+            'lastModTime' => 'DateUpdated',
+            'Total' => 'CountComments',
+            'Topics' => 'CountDiscussions',
+            'parent' => 'ParentCategoryID',
+        ];
         $ex->export(
             'Category',
             "select f.*
-                from :_Forums f;",
-            $category_Map
+                from :_Forums f
+                where linkTarget != 1", // External link categories have linkTarget==1
+            $map
         );
     }
 
@@ -131,21 +122,21 @@ class AspPlayground extends Source
      */
     protected function discussions(ExportModel $ex): void
     {
-        $discussion_Map = array(
+        $map = [
             'messageID' => 'DiscussionID',
             'ForumID' => 'CategoryID',
             'mem' => 'InsertUserID',
             'dateCreated' => 'DateInserted',
             'Subject' => 'Name',
             'hits' => 'CountViews',
-            'lastupdate' => 'DateLastComment'
-        );
+            'lastupdate' => 'DateLastComment',
+        ];
         $ex->export(
             'Discussion',
             "select t.*, m.Body
                 from :_Threads t
-                left join :_Messages m on m.messageID = t.messageID;",
-            $discussion_Map
+                left join :_Messages m on m.messageID = t.messageID",
+            $map
         );
     }
 
@@ -154,20 +145,54 @@ class AspPlayground extends Source
      */
     protected function comments(ExportModel $ex): void
     {
-        $comment_Map = array(
+        $map = [
             'messageID' => 'CommentID',
             'threadID' => 'DiscussionID',
-            'parent' => array('Column' => 'ReplyToCommentID', 'Type' => 'int'),
+            'parent' => 'ForeignID', // Preserve tree just in case.
             'Mem' => 'InsertUserID',
             'dateCreated' => 'DateInserted',
-            'Body' => 'Body',
-            'ip' => 'InsertIPAddress'
-        );
+            //'Body' => 'Body',
+        ];
+
+        // Avoid adding OP redundantly.
+        $skipOP = '';
+        if ($this->getDiscussionBodyMode()) {
+            $skipOP = "where parent != 0";
+        }
         $ex->export(
             'Comment',
             "select m.*, 'BBCode' as Format
-                from :_Messages m;",
-            $comment_Map
+                from :_Messages m
+                $skipOP",
+            $map
         );
+    }
+
+    /**
+     * @param ExportModel $ex
+     */
+    protected function bookmarks(ExportModel $ex): void
+    {
+        $map = [
+            'Mem' => 'UserID',
+            'threadID' => 'DiscussionID',
+        ];
+        $ex->export(
+            'UserDiscussion',
+            "select *, '1' as Bookmarked
+                from :_Subscription
+                where threadID is not null and isActive = 1",
+            $map
+        );
+    }
+
+    /**
+     * @param ExportModel $ex
+     */
+    protected function polls(ExportModel $ex): void
+    {
+        // Tables needed: Poll, PollDefinition, PollLog
+        // Not attached to discussions, only forums.
+        // Some allow multiple selections. Optional guest voting.
     }
 }
