@@ -1,7 +1,20 @@
 <?php
 
 // Autoload.
-require_once 'vendor/autoload.php';
+if (isset($GLOBALS['_composer_autoload_path'])) {
+    // If running via Composer, use provided location.
+    require_once $GLOBALS['_composer_autoload_path'];
+} else {
+    // If running locally, guess the location.
+    foreach (['../..', '../vendor', 'vendor'] as $path) {
+        $autoloader = __DIR__ . '/' . $path . '/autoload.php';
+        if (file_exists($autoloader)) {
+            require_once $autoloader;
+            unset($autoloader);
+            break;
+        }
+    }
+}
 
 // Environment.
 const ROOT_DIR = __DIR__;
@@ -12,6 +25,13 @@ if (ini_get('date.timezone') == '') {
 
 // Require & load config.
 \Porter\Config::getInstance()->set(loadConfig());
+
+// See deprecation notices in debug mode only.
+if (\Porter\Config::getInstance()->debugEnabled()) {
+    error_reporting(E_ALL);
+} else {
+    error_reporting(E_ALL & ~E_DEPRECATED);
+}
 
 // Load source & target support.
 \Porter\Support::getInstance()->setSources(loadSources());
