@@ -15,6 +15,7 @@ return [
     'FluxBb',
     'FuseTalk',
     'IpBoard3',
+    'IpBoard4',
     'JForum',
     'Kunena',
     'Mbox',
@@ -31,8 +32,8 @@ return [
     'Smf2',
     'Toast',
     'UserVoice',
+    'Vanilla',
     'Vanilla1',
-    'Vanilla2',
     'VBulletin',
     'VBulletin5',
     'WebWiz',
