@@ -2,11 +2,17 @@
 
 namespace Porter;
 
+use http\Exception;
+
 class Config
 {
     private static $instance = null;
 
-    protected array $config = [];
+    protected array $config = [
+        'debug' => false,
+        'test_alias' => 'test',
+        'connections' => [],
+    ];
 
     /**
      * Make it a singleton; there's only 1 config.
@@ -41,7 +47,7 @@ class Config
      */
     public function debugEnabled(): bool
     {
-        return $this->config['debug'];
+        return $this->config['debug'] ?? false;
     }
 
     /**
@@ -69,8 +75,42 @@ class Config
         foreach ($this->config['connections'] as $connection) {
             if ($alias === $connection['alias']) {
                 $result = $connection;
+                break;
             }
         }
+
+        $this->validateConnectionInfo($alias, $result);
+
         return $result;
+    }
+
+    /**
+     * Validate config has required info.
+     *
+     * @param string $alias
+     * @param array $info
+     * @throws \Exception
+     */
+    protected function validateConnectionInfo(string $alias, array $info): void
+    {
+        // Alias not in config.
+        if (empty($info)) {
+            //trigger_error('Config error: Alias "' . $alias . '" not found', E_USER_ERROR);
+            throw new \Exception('Alias "' . $alias . '" not found in config');
+        }
+
+        // Type is required.
+        if (empty($info['type'])) {
+            throw new \Exception('No connection `type` for alias "' . $alias . '" in config');
+        }
+
+        // Database required fields.
+        if ($info['type'] === 'database') {
+            foreach (['adapter', 'host','name','user'] as $property) {
+                if (!array_key_exists($property, $info)) {
+                    throw new \Exception('Database `' . $property . '` missing for alias "' . $alias . '" in config');
+                }
+            }
+        }
     }
 }
