@@ -1,11 +1,10 @@
 <?php
 
 /**
- * Invision Powerboard exporter tool.
+ * Invision Powerboard 3.x or earlier exporter tool.
  *
  * To export avatars, provide ?db-avatars=1&avatars-source=/path/to/avatars
  *
- * @license http://opensource.org/licenses/gpl-2.0.php GNU GPL2
  * @author  Lincoln Russell, lincolnwebs.com
  */
 
