@@ -106,7 +106,9 @@ abstract class Storage
             foreach ($row as $columnName => $value) {
                 if ($columnName === $src) {
                     $row[$dest] = $value; // Add column with new name.
-                    unset($row[$columnName]); // Remove old column.
+                    if ($dest !== $columnName) {
+                        unset($row[$columnName]); // Remove old column.
+                    }
                 }
             }
         }
@@ -143,16 +145,19 @@ abstract class Storage
     }
 
     /**
+     * Convert non-UTF-8 encodings to UTF-8.
+     *
      * @param array $row
      * @return array
      */
     public function fixEncoding(array $row): array
     {
         return array_map(function ($value) {
-            $doEncode = function_exists('mb_detect_encoding') &&
-                (mb_detect_encoding($value) != 'UTF-8') &&
+            $doEncode = $value && function_exists('mb_detect_encoding') &&
+                mb_detect_encoding($value) && // Verify we know the encoding at all.
+                (mb_detect_encoding($value) !== 'UTF-8') &&
                 (is_string($value) || is_numeric($value));
-            return ($doEncode) ? utf8_encode($value) : $value; // @todo Don't use utf8_encode()
+            return ($doEncode) ? mb_convert_encoding($value, 'UTF-8', mb_detect_encoding($value)) : $value;
         }, $row);
     }
 }
